LayoutTests/platform/mac-wk1/fast/scrolling/scroll-animator-basic-events-expected.txt

CONSOLE MESSAGE: line 25: RenderLayer: didAddVerticalScrollbar
CONSOLE MESSAGE: line 25: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 26: RenderLayer: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 26: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 27: RenderLayer: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 28: RenderLayer: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 28: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 29: RenderLayer: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 29: RenderLayer: mouseExitedContentArea
PASS successfullyParsed is true

TEST COMPLETE
This is a scrollable div.
